+++ src/pages/Leads/index.tsx
@@ -517,6 +517,8 @@ const Leads = () => {
             afterImportCompleted={() => {
               fetchLeads();
             }}
+            isExportAllOrSomeFeature={true}
+            total={rowCount}
             recordsToExport={selectedRecords.length}
             ids={selectedRecords.length ? selectedRecords.map((obj) => obj._id) : []}
             onExportToExcelSuccess={() => {
@@ -572,6 +574,7 @@ const Leads = () => {
               actionWidth={150}
               loading={loading}
               renderedFrom={leadResource}
+              refreshGrid={fetchLeads}
             /> : null
         }