+++ src/pages/Opportunities/index.tsx
@@ -483,6 +483,8 @@ const Opportunities = () => {
                   afterImportCompleted={() => {
                     fetchOpportunities();
                   }}
+                  isExportAllOrSomeFeature={true}
+                  total={rowCount}
                   recordsToExport={selectedRecords.length}
                   ids={selectedRecords.length ? selectedRecords.map((obj) => obj._id) : []}
                   onExportToExcelSuccess={() => {