+++ src/pages/RentalManagement/index.tsx
@@ -43,11 +43,11 @@ import { camelCase } from "lodash";
 let rentalManagementTimeout;
 const RentalManagementType = [
   {
-    key: "All Rental Managements",
+    key: `All ${routes.rentalManagement.title}`,
     value: 1,
   },
   {
-    key: "My Rental Managements",
+    key: `My ${routes.rentalManagement.title}`,
     value: 2,
   },
 ];
@@ -75,7 +75,7 @@ const RentalManagement = () => {
   const [singleRentalManagementDelete, setSingleRentalManagementDelete] = useState({
     id: null,
     show: false,
-    rentalManagementName: "",
+    rentalJobName: "",
   });
   const [accountDetails, setAccountDetails] = useState({
     accountId: history.location?.state?.accountId,
@@ -129,10 +129,7 @@ const RentalManagement = () => {
     let columns = []
     let rendererNames = []
     data.forEach(o => {
-      if (o?.fieldData?.fieldName === "rentalJobName") {
-        o.fieldData.primary = true
-      }
-      let currentColumn = getColumnData(pageTitle, o?.fieldData)
+      let currentColumn = getColumnData(pageTitle, o?.fieldData, routes.rentalManagementDetail.path)
       if (currentColumn !== null) {
         if (isOffline) {
           currentColumn.columnData["filter"] = false
@@ -210,7 +207,7 @@ const RentalManagement = () => {
         });
         fetchRentalManagement();
         dispatch({ type: "loading", loading: false });
-        setSingleRentalManagementDelete({ id: null, show: false, rentalManagementName: "" });
+        setSingleRentalManagementDelete({ id: null, show: false, rentalJobName: "" });
       })
       .catch((error) => {
         dispatch({ type: "loading", loading: false });
@@ -287,7 +284,7 @@ const RentalManagement = () => {
             setSingleRentalManagementDelete({
               show: true,
               id: params.data._id,
-              rentalManagementName: `${params.data.rentalManagementName}`,
+              rentalJobName: `${params.data.rentalJobName}`,
             })
           }
           entity="rentalManagement"
@@ -610,8 +607,7 @@ const RentalManagement = () => {
           {isConfirmDialogVisible ? (
             <ConfirmationDialog
               open={isConfirmDialogVisible}
-              message={`Are you sure you want to delete ${deleteRecord?.rentalManagementName ? "Rental Management" : "Rental Managements"
-                }   ${deleteRecord.rentalManagementName || ""}?`}
+              message={`Are you sure you want to delete selected ${routes.rentalManagement.title.toLowerCase()} ?`}
               onClose={() => {
                 if (deleteRecord) setDeleteRecord({});
                 setIsConformDialogVisible(false);
@@ -624,12 +620,12 @@ const RentalManagement = () => {
           {singleRentalManagementDelete.show ? (
             <ConfirmationDialog
               open={singleRentalManagementDelete.show}
-              message={`Are you sure you want to delete Rental Management: ${singleRentalManagementDelete.rentalManagementName}?`}
+              message={`Are you sure you want to delete this ${routes.rentalManagement.title.toLowerCase()} ${singleRentalManagementDelete ? singleRentalManagementDelete?.id ? singleRentalManagementDelete?.rentalJobName : "" : ""}?`}
               onClose={() =>
                 setSingleRentalManagementDelete({
                   id: null,
                   show: false,
-                  rentalManagementName: "",
+                  rentalJobName: "",
                 })
               }
               onOk={handleSingleDeleteRentalManagement}